fix: untangle dependency cycle in subsidy request contexts/hooks.
