+++ src/components/enterprise-subsidy-requests/data/hooks.js
@@ -1,12 +1,11 @@
 import {
-  useState, useEffect, useContext, useMemo,
+  useState, useEffect,
 } from 'react';
 import { logError } from '@edx/frontend-platform/logging';
 import { camelCaseObject } from '@edx/frontend-platform/utils';
 import { getAuthenticatedUser } from '@edx/frontend-platform/auth';
 import { fetchSubsidyRequestConfiguration, fetchLicenseRequests, fetchCouponCodeRequests } from './service';
 import { SUBSIDY_TYPE, SUBSIDY_REQUEST_STATE } from '../constants';
-import { SubsidyRequestsContext } from '../SubsidyRequestsContextProvider';
 
 export function useSubsidyRequestConfiguration(enterpriseUUID) {
   const [subsidyRequestConfiguration, setSubsidyRequestConfiguration] = useState();
@@ -101,47 +100,4 @@ export function useSubsidyRequests(subsidyRequestConfiguration) {
     isLoading,
     refreshSubsidyRequests: loadSubsidyRequests,
   };
-}
-
-/**
- * Returns `true` if user has made a subsidy request.
- *
- * Returns `false` if:
- *  - Subsidy request has not been configured
- *  - No requests are found under the configured `SUBSIDY_TYPE`
- *
- * If the `SUBSIDY_TYPE` is `COUPON`, optional parameter courseKey can be passed
- * to only return true if courseKey is in one of the requests
- *
- * @param {string} [courseKey] - optional filter for specific course
- * @returns {boolean}
- */
-export function useUserHasSubsidyRequestForCourse(courseKey) {
-  const {
-    subsidyRequestConfiguration,
-    requestsBySubsidyType,
-  } = useContext(SubsidyRequestsContext);
-
-  return useMemo(() => {
-    if (!subsidyRequestConfiguration?.subsidyRequestsEnabled) {
-      return false;
-    }
-    switch (subsidyRequestConfiguration.subsidyType) {
-      case SUBSIDY_TYPE.LICENSE: {
-        return requestsBySubsidyType[SUBSIDY_TYPE.LICENSE].length > 0;
-      }
-      case SUBSIDY_TYPE.COUPON: {
-        const foundCouponRequest = requestsBySubsidyType[SUBSIDY_TYPE.COUPON].find(
-          request => (!courseKey || request.courseId === courseKey),
-        );
-        return !!foundCouponRequest;
-      }
-      default:
-        return false;
-    }
-  }, [
-    courseKey,
-    subsidyRequestConfiguration,
-    requestsBySubsidyType,
-  ]);
 }
